Automated Test Grader
Reads a txt fstream file that contains a 5 letter key to the answers for a test. Along with the answers and names of participating students. The program then runs through the text file to determine to average scores of all participants as well as prints the results of the test as a score out of 5. The display is formatted with std::setw().